fix instrumentation for Rails 7 ...
- fix test setup
- remove more older Rails stuff

NH-5971
