+++ .travis.yml
@@ -28,7 +28,7 @@ gemfile:
   - gemfiles/instrumentation_mocked.gemfile
   - gemfiles/instrumentation_mocked_oldgems.gemfile
   - gemfiles/frameworks.gemfile
-  #  - gemfiles/rails70.gemfile
+  - gemfiles/rails70.gemfile
   - gemfiles/rails61.gemfile
   - gemfiles/rails52.gemfile
   - gemfiles/delayed_job.gemfile
