+++ Rakefile
@@ -31,19 +31,8 @@ Rake::TestTask.new do |t|
     require 'delayed/tasks'
     t.test_files = FileList['test/queues/delayed_job*_test.rb']
   when /rails/
-    # Pre-load rails to get the major version number
-    require 'rails'
-
-    if Rails::VERSION::MAJOR == 5
-      t.test_files = FileList["test/frameworks/rails#{Rails::VERSION::MAJOR}x_test.rb"] +
-                     FileList["test/frameworks/rails#{Rails::VERSION::MAJOR}x_api_test.rb"]
-    elsif Rails::VERSION::MAJOR >= 6
-      t.test_files = FileList['test/frameworks/rails5x_test.rb'] +
-                     FileList['test/frameworks/rails5x_api_test.rb']
-    else
-      t.test_files = FileList["test/frameworks/rails#{Rails::VERSION::MAJOR}x_test.rb"]
-    end
-
+    t.test_files = FileList['test/frameworks/rails5x_test.rb'] +
+                   FileList['test/frameworks/rails5x_api_test.rb']
   when /frameworks/
     t.test_files = FileList['test/frameworks/sinatra*_test.rb'] +
                    FileList['test/frameworks/padrino*_test.rb'] +
