+++ gemfiles/rails61_api.gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-gem 'rails', '~> 6.0.0'
+gem 'rails', '~> 6.1.0'
 gem 'sass-rails'
 gem 'uglifier'
 gem 'coffee-rails'
