+++ gemfiles/rails70_api.gemfile
@@ -1,0 +1,38 @@
+source "https://rubygems.org"
+
+gem 'rails', '~> 7.0.0'
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
+# gem 'therubyracer', platforms: :ruby # not maintained anymore
+gem 'jquery-rails'
+gem 'turbolinks'
+gem 'jbuilder'
+gem 'sidekiq'
+gem 'lograge'
+gem 'wicked_pdf'
+gem 'wkhtmltopdf-binary'
+
+group :development do
+  # Access an IRB console on exception pages or by using <%= console %> in views
+  gem 'web-console'
+  gem 'listen'
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+  gem 'spring-watcher-listen'
+end
+
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
+
+if defined?(JRUBY_VERSION)
+  gem 'activerecord-jdbc-adapter'
+  gem 'jdbc-postgresql'
+else
+  gem 'pg'
+  gem 'mysql2'
+end
+
+eval(File.read(File.join(File.dirname(__FILE__), 'test_gems.gemfile')))
+gemspec :path => File.expand_path(File.dirname(__FILE__) + '/../')
+# vim:syntax=ruby
