+++ lib/appoptics_apm/frameworks/rails/inst/action_controller.rb
@@ -84,18 +84,14 @@ module AppOpticsAPM
 end
 
 # ActionController::Base
-if defined?(ActionController::Base) && AppOpticsAPM::Config[:action_controller][:enabled] && Rails::VERSION::MAJOR <= 6
+if defined?(ActionController::Base) && AppOpticsAPM::Config[:action_controller][:enabled]
   AppOpticsAPM.logger.info '[appoptics_apm/loading] Instrumenting actioncontroller' if AppOpticsAPM::Config[:verbose]
-  require "appoptics_apm/frameworks/rails/inst/action_controller#{Rails::VERSION::MAJOR}"
-  if Rails::VERSION::MAJOR >= 5
-    ActionController::Base.send(:prepend, ::AppOpticsAPM::Inst::ActionController)
-  elsif Rails::VERSION::MAJOR < 5
-    AppOpticsAPM::Util.send_include(::ActionController::Base, AppOpticsAPM::Inst::ActionController)
-  end
+  require "appoptics_apm/frameworks/rails/inst/action_controller5"
+  ActionController::Base.send(:prepend, ::AppOpticsAPM::Inst::ActionController)
 end
 
-# ActionController::API - Rails 5 or via the rails-api gem
-if defined?(ActionController::API) && AppOpticsAPM::Config[:action_controller_api][:enabled]  && Rails::VERSION::MAJOR <= 6
+# ActionController::API
+if defined?(ActionController::API) && AppOpticsAPM::Config[:action_controller_api][:enabled]
   AppOpticsAPM.logger.info '[appoptics_apm/loading] Instrumenting actioncontroller api' if AppOpticsAPM::Config[:verbose]
   require "appoptics_apm/frameworks/rails/inst/action_controller_api"
   ActionController::API.send(:prepend, ::AppOpticsAPM::Inst::ActionControllerAPI)
