+++ test/run_tests/.ruby_version_alpine
@@ -1,1 +1,1 @@
-2.5.9
+3.0.3
